+++ langgraph_agent_toolkit/agents/blueprints/chatbot/agent.py
@@ -20,11 +20,11 @@ async def chatbot(
 ):
     messages = inputs["messages"]
     if previous:
-        messages = add_messages(previous["messages"], messages)
+        messages = previous["messages"] + messages
 
-    model = ModelFactory.create(config["configurable"].get("model", settings.DEFAULT_MODEL))
+    model = ModelFactory.create(config["configurable"].get("model", settings.DEFAULT_MODEL_TYPE))
     response = await model.ainvoke(messages)
-    return entrypoint.final(value={"messages": [response]}, save={"messages": add_messages(messages, response)})
+    return entrypoint.final(value={"messages": [response]}, save={"messages": messages + [response]})
 
 
 chatbot_agent = Agent(