+++ langgraph_agent_toolkit/helper/constants.py
@@ -15,10 +15,12 @@ def set_default_agent(agent_name):
 
 DEFAULT_MAX_MESSAGE_HISTORY_LENGTH = 6 + 1  # N messages + 1 system message
 DEFAULT_RECURSION_LIMIT = 25
-DEFAULT_OPENAI_COMPATIBLE_MODEL_PARAMS = dict(
+DEFAULT_OPENAI_MODEL_TYPE_PARAMS = dict(
     temperature=0.0,
     max_tokens=1024,
     top_p=0.7,
     streaming=True,
 )
 DEFAULT_CACHE_TTL_SECOND = 60 * 10  # 10 minutes
+
+DEFAULT_STREAMLIT_USER_ID = "streamlit-user"