+++ langgraph_agent_toolkit/schema/schema.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel, Field, SerializeAsAny
 from typing_extensions import TypedDict
 
 from langgraph_agent_toolkit.helper.constants import (
-    DEFAULT_OPENAI_COMPATIBLE_MODEL_PARAMS,
+    DEFAULT_OPENAI_MODEL_TYPE_PARAMS,
     DEFAULT_RECURSION_LIMIT,
     get_default_agent,
 )
@@ -37,7 +37,7 @@ class ServiceMetadata(BaseModel):
         description="Default agent used when none is specified.",
         examples=[get_default_agent()],
     )
-    default_model: AllModelEnum = Field(
+    default_model_type: AllModelEnum = Field(
         description="Default model used when none is specified.",
     )
 
@@ -60,13 +60,18 @@ class UserInput(BaseModel):
         default=None,
         examples=["847c6285-8fc9-4560-a83f-4e6285809254"],
     )
+    user_id: str | None = Field(
+        description="User ID to persist in observability platform and share long-term memory.",
+        default=None,
+        examples=["521c0a60-ea75-43fa-a793-a4cf11e013ae"],
+    )
     agent_config: dict[str, Any] = Field(
         description="Additional configuration to pass through to the agent",
         default={},
         examples=[
             {
                 "memory_saver_params": {"k": 6},
-                **DEFAULT_OPENAI_COMPATIBLE_MODEL_PARAMS,
+                **DEFAULT_OPENAI_MODEL_TYPE_PARAMS,
             },
         ],
     )
@@ -161,6 +166,11 @@ class Feedback(BaseModel):
         description="Feedback score.",
         examples=[0.8],
     )
+    user_id: str | None = Field(
+        description="User ID to associate with the feedback.",
+        default=None,
+        examples=["521c0a60-ea75-43fa-a793-a4cf11e013ae"],
+    )
     kwargs: dict[str, Any] = Field(
         description="Additional feedback kwargs, passed to LangSmith.",
         default={},
@@ -188,6 +198,11 @@ class ChatHistoryInput(BaseModel):
     thread_id: str = Field(
         description="Thread ID to persist and continue a multi-turn conversation.",
         examples=["847c6285-8fc9-4560-a83f-4e6285809254"],
+    )
+    user_id: str | None = Field(
+        description="User ID to persist in observability platform and share long-term memory.",
+        default=None,
+        examples=["521c0a60-ea75-43fa-a793-a4cf11e013ae"],
     )