+++ langgraph_agent_toolkit/service/routes.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, HTTPException, Request, status
 from fastapi.responses import RedirectResponse, StreamingResponse
 from langchain_core.messages import AnyMessage
 from langchain_core.runnables import RunnableConfig
-from langgraph.graph.state import CompiledStateGraph
+from langgraph.func import Pregel
 
 from langgraph_agent_toolkit import __version__
 from langgraph_agent_toolkit.agents.agent import Agent
@@ -43,7 +43,7 @@ async def info(request: Request) -> ServiceMetadata:
         agents=get_all_agent_info(request),
         models=models,
         default_agent=get_default_agent(),
-        default_model=settings.DEFAULT_MODEL,
+        default_model_type=settings.DEFAULT_MODEL_TYPE,
     )
 
 
@@ -66,6 +66,7 @@ async def invoke(user_input: UserInput, agent_id: str = None, request: Request =
             agent_id=agent_id,
             message=user_input.message,
             thread_id=user_input.thread_id,
+            user_id=user_input.user_id,
             model=user_input.model,
             agent_config=user_input.agent_config,
             recursion_limit=user_input.recursion_limit,
@@ -119,6 +120,7 @@ async def feedback(
             run_id=feedback.run_id,
             key=feedback.key,
             score=feedback.score,
+            user_id=feedback.user_id,
             **feedback.kwargs,
         )
 
@@ -140,11 +142,12 @@ def history(input: ChatHistoryInput, request: Request = None) -> ChatHistory:
     """Get chat history."""
     agent: Agent = get_agent(request, get_default_agent())
     try:
-        agent_graph: CompiledStateGraph = agent.graph
+        agent_graph: Pregel = agent.graph
         state_snapshot = agent_graph.get_state(
             config=RunnableConfig(
                 configurable={
                     "thread_id": input.thread_id,
+                    "user_id": input.user_id,
                 }
             )
         )