+++ CHANGELOG.md
@@ -6,6 +6,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2]
+
+### Added
+
+- `user_id` parameter
+- `store` creator to memory classes
+
+### Fixed
+
+- enhance error handling and testing in Streamlit app
+- add new chat button
+- variable names
+- type hints
+
+### Removed
+
+- print statements
+
 ## [0.1.1]
 
 ### Added